Updated mileage page content.
